+++ src/main/java/com/github/karlnicholas/hdf5javalib/datatype/HdfDataType.java
@@ -1,5 +1,5 @@
 package com.github.karlnicholas.hdf5javalib.datatype;
 
 public interface HdfDataType {
-    public int getSizeMessageData();
+    public short getSizeMessageData();
 }
