+++ src/main/java/com/github/karlnicholas/hdf5javalib/datatype/HdfString.java
@@ -92,8 +92,8 @@ public class HdfString implements HdfDataType {
     }
 
     @Override
-    public int getSizeMessageData() {
-        return bytes.length;
+    public short getSizeMessageData() {
+        return (short)bytes.length;
     }
 
 //    @Override
