+++ src/main/java/com/github/karlnicholas/hdf5javalib/message/BTreeKValuesMessage.java
@@ -19,7 +19,7 @@ public class BTreeKValuesMessage extends HdfMessage {
     private int groupLeafNodeK;
 
     public BTreeKValuesMessage(int version, int indexedStorageInternalNodeK, int groupInternalNodeK, int groupLeafNodeK) {
-        super(19, ()->1+2+2+2, (byte)0);
+        super((short)19, ()-> (short) (1+2+2+2), (byte)0);
         this.version = version;
         this.indexedStorageInternalNodeK = indexedStorageInternalNodeK;
         this.groupInternalNodeK = groupInternalNodeK;
